resin-init-flasher.bbappend: Use new flasher variables
This .bbappend needs to stay aligned with the changes done to resin-init-flasher
in meta-resin. These changes involve removing board specific bootloader variables
and instead now we just specify the name of the file containing bootloader config
options for when booting from internal media along with the pathname where this
file should be copied over by the flasher to the internal media.

Signed-off-by: Florin Sarbu <florin@resin.io>