u-boot-ti-staging: uEnv.txt_internal functionality was replaced
The old functionality of uEnv.txt was replaced by a special resin
configuration file for the bootloader. The old file though is still
useful as the BSP takes advantage of it to load for example a specific
dtb. Keep it empty so users can write custom configuration which will
end up on the internal device.

Signed-off-by: Andrei Gherzan <andrei@resin.io>