u-boot-ti-staging: Adapt to new codebase from rocko BSP layer
We updated the BSP layer to rocko version. This brings an updated
codebase for the bootloader and we need to adapt to it, namely: remove
the obsolete patches that added support for BBBW, BBGW, BBBlue since
those patches are merged upstream and also rework the Balena u-boot
integration patch to work with the updated u-boot code.

Signed-off-by: Florin Sarbu <florin@balena.io>