u-boot-ti-staging: Fix SPL boot loop
The BSP layer at the rocko branch has an issue with the boot
stuck in a loop. See details here:
https://lists.yoctoproject.org/pipermail/meta-ti/2018-October
/011363.html
Also, when using the am335x_evm_config defconfig, the resulted
SPL is bigger than the available SRAM so we also need to shrink
the SPL size.

Changelog-entry: Fix boot loop hang when switched to meta-ti rocko
Signed-off-by: Florin Sarbu <florin@balena.io>