Add scripts for PocketBeagle usb ethernet
Integrated boot scripts used by the BeagleBone
debian images.

Adapted these to work with Balena and provide
out-of-the-box networking for PocketBeagle.

Internet connection sharing needs to be used
with the network created by default when
plugging the PocketBeagle to a host using the
on-board microUSB port.

Changelog-entry: Add scripts for PocketBeagle networking
Signed-off-by: Alexandru Costache <alexandru@balena.io>