beaglebone green wifi: Add back notice for boot button
According to the upstream documentation, by default
the beaglebone green will load uboot from sd-card, and switch
to emmc if boot button is pressed.

However, this statement is incorrect and the issue
is masked if BalenaOS is already present on the emmc.
With an official debian image on the eMMC, it will be
necessary to power the board while holding down
the USER/BOOT button.

Signed-off-by: Alexandru Costache <alexandru@balena.io>