linux-beagleboard: Update to 5.4.47-ti-r13
The kernel defconfig is the arch/arm/configs/bb.org_defconfig file
from the kernel source tree at tag 5.4.47-ti-r13.

Removed 0002-rtc-omap-Prevent-kernel-panic-and-reboot-on-shutdown.patch
since the logic it was patching has changed in this newer kernel.

Removed the Netflix vuln patches since they are merged in the 5.4.47
kernel.

Changelog-entry: Update kernel for the beaglebones to version 5.4.47-ti-r13
Signed-off-by: Florin Sarbu <florin@balena.io>