meta-balena-beaglebone: local.conf.sample: Modify hosttools
Using a rootless balena engine to build removes the docker dependency
but introduces a few other dependencies.

newuimap and newgidmap can be installed with the shadow package but they
need to be setuid.

The `ip` and `ps` utilities do not currently have a native recipe and
they are common enough not to be a problem.

Until `aufs` is deprecated the docker dependency cannot be removed.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>