linux-beagleboard: Unset CONFIG_DEFAULT_SECURITY_APPARMOR in defconfig
We need to have apparmor removed, for details see
https://github.com/balena-os/meta-balena/commit/18cd233a83554b58b3540164afd768fdeda60b03

However, due to the nature of this kernel setting, if it's enabled in the BSP then the above
commit won't disable it in our case. So instead we just modify the specific kernel recipe to
directly remove apparmor support from here.

Changelog-entry: Unset CONFIG_DEFAULT_SECURITY_APPARMOR in kernel defconfig
Signed-off-by: Florin Sarbu <florin@balena.io>