Update to kirkstone.
In order to make the beaglebone-ai64 to work it is first required to
switch to meta-ti's kirkstone branch, so we need to update all the
relevant layers.

This also includes a fix for the gen_mod_headers script, as it seems
Balena is passing a wrong 3rd argument. Fix by Jon Trulson
<jtrulson@ics.com>

We also need to manually remove kernel-related image/headers from
MACHINE_ESSENTIAL_EXTRA_RDEPENDS.

On kirkstone meta-ti uses linux-bb.org, so we created a bbappend in
order for it to inherit kernel-balena. This kernel might probably be
used with the beaglebone AI64.

The kernel-modules-headers' bbappend is removed because the base recipe
does not exists anymore on any layer.

Changelog-entry: Update layer to kirkstone
Signed-off-by: Lisandro Pérez Meyer <lperezmeyer@ics.com>