BBAI64: Add support for the correct resin partitions.
It is worth to note that, using the current meta-ti version in this
repository,u-boot-ti-staging seems to be using j721e_evm_a72_defconfig
instead of the j721e_beagleboneai64_a72.config file.

This works, but an update of meta-ti and the subsequent fixes should be
considered.

Changed-by: Lisandro Pérez Meyer <lpmeyer@ics.com>
Changelog-Entry: bbai64: use the correct resin partitions when booting