BBAI64: Try to use u-boot-bb.org
The currrent u-boot-ti-staging recipe is too old, so switching to what
we know it works: u-boot-bb.org.

This also requires adding two new recipes coming from poky which are not
yet on kirkstone.

Clear TI_TMPDIR_APPEND so that the files for both archs are in the same
path and we can install them.

Change the path to where the device trees are, thereis not ti
subdirectory.