linux-firmware: Update overrides for firmware compression
At this moment we override a few firmware packages to only ship
the latest version to save space. These overrides hardcode file names,
which won't match when firmware compression is enabled and a suffix
is appended.

This patch appends a wildcard to the file names which works
both with & without suffix/compression.

Changelog-entry: Enable firmware compression
Signed-off-by: Michal Toman <michalt@balena.io>